backend-fastapi-aiagent/requirements.txt
fastapi
uvicorn
httpx
python-dotenv
openai>=1.0.0